+++ src/aiohttp_sqlalchemy/typedefs.py
@@ -1,12 +1,11 @@
 from typing import Awaitable, Callable, Iterable, Tuple, Union
 
 from aiohttp.web import StreamResponse
-from sqlalchemy.ext.asyncio import AsyncEngine
-from sqlalchemy.orm import Session, sessionmaker
+from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, async_sessionmaker
 
 THandler = Callable[..., Awaitable[StreamResponse]]
 THandlerWrapper = Callable[..., THandler]
 
-TTarget = Union[str, AsyncEngine, sessionmaker[Session]]
-TBind = Tuple[sessionmaker[Session], str, bool]
+TTarget = Union[str, AsyncEngine, async_sessionmaker[AsyncSession]]
+TBind = Tuple[async_sessionmaker[AsyncSession], str, bool]
 TBinds = Iterable[TBind]